[Refactor] Extract QueryRuntimeProfile
