+++ fe/fe-core/src/main/java/com/starrocks/qe/DefaultCoordinator.java
@@ -34,7 +34,6 @@
 
 package com.starrocks.qe;
 
-import com.google.common.base.Preconditions;
 import com.google.common.base.Strings;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
@@ -44,18 +43,12 @@ import com.starrocks.authentication.AuthenticationMgr;
 import com.starrocks.catalog.FsBroker;
 import com.starrocks.common.Config;
 import com.starrocks.common.FeConstants;
-import com.starrocks.common.MarkedCountDownLatch;
-import com.starrocks.common.Pair;
 import com.starrocks.common.Status;
 import com.starrocks.common.ThriftServer;
 import com.starrocks.common.UserException;
-import com.starrocks.common.util.Counter;
 import com.starrocks.common.util.DebugUtil;
-import com.starrocks.common.util.ProfileManager;
-import com.starrocks.common.util.ProfilingExecPlan;
 import com.starrocks.common.util.RuntimeProfile;
 import com.starrocks.connector.exception.RemoteFileNotFoundException;
-import com.starrocks.load.loadv2.LoadJob;
 import com.starrocks.planner.PlanFragment;
 import com.starrocks.planner.ResultSink;
 import com.starrocks.planner.RuntimeFilterDescription;
@@ -65,6 +58,7 @@ import com.starrocks.privilege.PrivilegeBuiltinConstants;
 import com.starrocks.proto.PPlanFragmentCancelReason;
 import com.starrocks.qe.scheduler.Coordinator;
 import com.starrocks.qe.scheduler.Deployer;
+import com.starrocks.qe.scheduler.QueryRuntimeProfile;
 import com.starrocks.qe.scheduler.dag.ExecutionDAG;
 import com.starrocks.qe.scheduler.dag.ExecutionFragment;
 import com.starrocks.qe.scheduler.dag.FragmentInstance;
@@ -78,11 +72,9 @@ import com.starrocks.sql.PlannerProfile;
 import com.starrocks.sql.ast.UserIdentity;
 import com.starrocks.sql.plan.ExecPlan;
 import com.starrocks.system.ComputeNode;
-import com.starrocks.task.LoadEtlTask;
 import com.starrocks.thrift.TDescriptorTable;
 import com.starrocks.thrift.TLoadJobType;
 import com.starrocks.thrift.TNetworkAddress;
-import com.starrocks.thrift.TPipelineProfileLevel;
 import com.starrocks.thrift.TQueryType;
 import com.starrocks.thrift.TReportExecStatusParams;
 import com.starrocks.thrift.TRuntimeFilterDestination;
@@ -92,21 +84,18 @@ import com.starrocks.thrift.TStatusCode;
 import com.starrocks.thrift.TTabletCommitInfo;
 import com.starrocks.thrift.TTabletFailInfo;
 import com.starrocks.thrift.TUniqueId;
-import com.starrocks.thrift.TUnit;
-import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 import java.util.function.Supplier;
@@ -117,58 +106,38 @@ public class DefaultCoordinator extends Coordinator {
 
     private static final int DEFAULT_PROFILE_TIMEOUT_SECOND = 2;
 
-    // Overall status of the entire query; set to the first reported fragment error
-    // status or to CANCELLED, if Cancel() is called.
-    Status queryStatus = new Status();
-
     private final JobSpec jobSpec;
     private final ExecutionDAG executionDAG;
 
-    // protects all fields below
-    private final Lock lock = new ReentrantLock();
-    // If true, the query is done returning all results.  It is possible that the
-    // coordinator still needs to wait for cleanup on remote fragments (e.g. queries
-    // with limit)
-    // Once this is set to true, errors from remote fragments are ignored.
-    private boolean returnedAllResults;
-    private RuntimeProfile queryProfile;
-    private List<RuntimeProfile> fragmentProfiles;
-    private ResultReceiver receiver;
-    // number of instances of this query, equals to
-    // number of backends executing plan fragments on behalf of this query;
-    // set in computeFragmentExecParams();
-    // same as backend_exec_states_.size() after Exec()
-    // instance id -> dummy value
-    private MarkedCountDownLatch<TUniqueId, Long> profileDoneSignal;
-    private int numReceivedRows = 0;
-    private List<String> deltaUrls;
-    private Map<String, String> loadCounters;
-    private String trackingUrl;
-    private final Set<String> rejectedRecordPaths = new HashSet<>();
-    // for export
-    private List<String> exportFiles;
-    private final List<TTabletCommitInfo> commitInfos = Lists.newArrayList();
-    private final List<TTabletFailInfo> failInfos = Lists.newArrayList();
-
-    // for external table sink
-    private final List<TSinkCommitInfo> sinkCommitInfos = Lists.newArrayList();
-    // Input parameter
     private final ConnectContext connectContext;
     private final boolean needReport;
 
-    // True indicates that the profile has been reported
-    // When `enable_load_profile` is enabled,
-    // if the time costs of stream load is less than `stream_load_profile_collect_second`,
-    // the profile will not be reported to FE to reduce the overhead of profile under high-frequency import
-    private boolean profileAlreadyReported = false;
-
     private final CoordinatorPreprocessor coordinatorPreprocessor;
 
-    private boolean thriftServerHighLoad;
+    /**
+     * Protects all the fields below.
+     */
+    private final Lock lock = new ReentrantLock();
 
-    private Supplier<RuntimeProfile> topProfileSupplier;
-    private Supplier<ExecPlan> execPlanSupplier;
-    private final AtomicLong lastRuntimeProfileUpdateTime = new AtomicLong(System.currentTimeMillis());
+    /**
+     * Overall status of the entire query.
+     * <p> Set to the first reported fragment error status or to CANCELLED, if {@link #cancel()} is called.
+     */
+    private Status queryStatus = new Status();
+
+    private final QueryRuntimeProfile queryProfile;
+
+    private ResultReceiver receiver;
+    private int numReceivedRows = 0;
+
+    /**
+     * True indicates the query is done returning all results.
+     * <p> It is possible that the coordinator still needs to wait for cleanup on remote fragments (e.g. queries with limit)
+     * Once this is set to true, errors from remote fragments are ignored.
+     */
+    private boolean returnedAllResults;
+
+    private boolean thriftServerHighLoad;
 
     private LogicalSlot slot = null;
 
@@ -251,8 +220,7 @@ public class DefaultCoordinator extends Coordinator {
     }
 
     /**
-     * Only used for sync stream load profile,
-     * so only init relative data structure.
+     * Only used for sync stream load profile, and only init relative data structure.
      */
     public DefaultCoordinator(JobSpec jobSpec, StreamLoadPlanner planner, TNetworkAddress address) {
         this.connectContext = planner.getConnectContext();
@@ -262,21 +230,13 @@ public class DefaultCoordinator extends Coordinator {
         TUniqueId queryId = jobSpec.getQueryId();
 
         LOG.info("Execution Profile: {}", DebugUtil.printId(queryId));
-        queryProfile = new RuntimeProfile("Execution");
-
-        fragmentProfiles = new ArrayList<>();
-        fragmentProfiles.add(new RuntimeProfile("Fragment 0"));
-        queryProfile.addChild(fragmentProfiles.get(0));
-        profileDoneSignal = new MarkedCountDownLatch<>(1);
-        profileDoneSignal.addMark(queryId, -1L /* value is meaningless */);
 
         FragmentInstanceExecState execState = FragmentInstanceExecState.createFakeExecution(queryId, address);
         executionDAG.addExecution(execState);
 
-        attachInstanceProfileToFragmentProfile();
-
-        deltaUrls = Lists.newArrayList();
-        loadCounters = Maps.newHashMap();
+        this.queryProfile = new QueryRuntimeProfile(connectContext, jobSpec, 1);
+        queryProfile.attachInstances(Collections.singletonList(queryId));
+        queryProfile.attachExecutionProfiles(executionDAG.getExecutions());
 
         this.coordinatorPreprocessor = null;
         this.needReport = true;
@@ -290,6 +250,8 @@ public class DefaultCoordinator extends Coordinator {
 
         this.coordinatorPreprocessor = new CoordinatorPreprocessor(context, jobSpec);
         this.executionDAG = coordinatorPreprocessor.getExecutionDAG();
+
+        this.queryProfile = new QueryRuntimeProfile(connectContext, jobSpec, executionDAG.getFragmentsInCreatedOrder().size());
     }
 
     @Override
@@ -333,27 +295,27 @@ public class DefaultCoordinator extends Coordinator {
 
     @Override
     public RuntimeProfile getQueryProfile() {
-        return queryProfile;
+        return queryProfile.getQueryProfile();
     }
 
     @Override
     public List<String> getDeltaUrls() {
-        return deltaUrls;
+        return queryProfile.getDeltaUrls();
     }
 
     @Override
     public Map<String, String> getLoadCounters() {
-        return loadCounters;
+        return queryProfile.getLoadCounters();
     }
 
     @Override
     public String getTrackingUrl() {
-        return trackingUrl;
+        return queryProfile.getTrackingUrl();
     }
 
     @Override
     public List<String> getRejectedRecordPaths() {
-        return new ArrayList<>(rejectedRecordPaths);
+        return queryProfile.getRejectedRecordPaths();
     }
 
     @Override
@@ -375,11 +337,8 @@ public class DefaultCoordinator extends Coordinator {
         lock.lock();
         try {
             executionDAG.resetExecutions();
-            this.queryStatus.setStatus(new Status());
-            if (this.exportFiles == null) {
-                this.exportFiles = Lists.newArrayList();
-            }
-            this.exportFiles.clear();
+            queryStatus.setStatus(new Status());
+            queryProfile.clearExportStatus();
         } finally {
             lock.unlock();
         }
@@ -387,27 +346,32 @@ public class DefaultCoordinator extends Coordinator {
 
     @Override
     public List<TTabletCommitInfo> getCommitInfos() {
-        return commitInfos;
+        return queryProfile.getCommitInfos();
     }
 
     @Override
     public List<TTabletFailInfo> getFailInfos() {
-        return failInfos;
+        return queryProfile.getFailInfos();
     }
 
     @Override
     public List<TSinkCommitInfo> getSinkCommitInfos() {
-        return sinkCommitInfos;
+        return queryProfile.getSinkCommitInfos();
+    }
+
+    @Override
+    public List<String> getExportFiles() {
+        return queryProfile.getExportFiles();
     }
 
     @Override
     public void setTopProfileSupplier(Supplier<RuntimeProfile> topProfileSupplier) {
-        this.topProfileSupplier = topProfileSupplier;
+        queryProfile.setTopProfileSupplier(topProfileSupplier);
     }
 
     @Override
     public void setExecPlanSupplier(Supplier<ExecPlan> execPlanSupplier) {
-        this.execPlanSupplier = execPlanSupplier;
+        queryProfile.setExecPlanSupplier(execPlanSupplier);
     }
 
     @Override
@@ -450,7 +414,6 @@ public class DefaultCoordinator extends Coordinator {
 
         coordinatorPreprocessor.prepareExec();
 
-        // create result receiver
         prepareResultSink();
 
         prepareProfile();
@@ -492,6 +455,11 @@ public class DefaultCoordinator extends Coordinator {
         try (PlannerProfile.ScopedTimer timer = PlannerProfile.getScopedTimer("Scheduler.Deploy")) {
             deliverExecFragments(needDeploy);
         }
+
+        // Prevent `explain scheduler` from waiting until the profile timeout.
+        if (!needDeploy) {
+            queryProfile.finishAllInstances(Status.OK);
+        }
     }
 
     @Override
@@ -502,62 +470,50 @@ public class DefaultCoordinator extends Coordinator {
     }
 
     private void prepareProfile() {
-        queryProfile = new RuntimeProfile("Execution");
-        fragmentProfiles = new ArrayList<>();
-        for (ExecutionFragment fragment : executionDAG.getFragmentsInCreatedOrder()) {
-            RuntimeProfile fragmentProfile = new RuntimeProfile("Fragment " + fragment.getFragmentIndex());
-            fragmentProfiles.add(fragmentProfile);
-            queryProfile.addChild(fragmentProfile);
+        ExecutionFragment rootExecFragment = executionDAG.getRootFragment();
+        boolean isLoadType = !(rootExecFragment.getPlanFragment().getSink() instanceof ResultSink);
+        if (isLoadType) {
+            jobSpec.getQueryOptions().setEnable_profile(true);
+            List<Long> relatedBackendIds = coordinatorPreprocessor.getWorkerProvider().getSelectedWorkerIds();
+            GlobalStateMgr.getCurrentState().getLoadMgr().initJobProgress(
+                    jobSpec.getLoadJobId(), jobSpec.getQueryId(), executionDAG.getInstanceIds(), relatedBackendIds);
+            LOG.info("dispatch load job: {} to {}", DebugUtil.printId(jobSpec.getQueryId()),
+                    coordinatorPreprocessor.getWorkerProvider().getSelectedWorkerIds());
         }
 
-        // to keep things simple, make async Cancel() calls wait until plan fragment
-        // execState has been initiated, otherwise we might try to cancel fragment
-        // execState at backends where it hasn't even started
-        profileDoneSignal = new MarkedCountDownLatch<>(executionDAG.getInstanceIds().size());
-        for (TUniqueId instanceId : executionDAG.getInstanceIds()) {
-            profileDoneSignal.addMark(instanceId, -1L /* value is meaningless */);
-        }
+        queryProfile.attachInstances(executionDAG.getInstanceIds());
     }
 
     private void prepareResultSink() throws Exception {
         ExecutionFragment rootExecFragment = executionDAG.getRootFragment();
-        if (rootExecFragment.getPlanFragment().getSink() instanceof ResultSink) {
-            long workerId = rootExecFragment.getInstances().get(0).getWorkerId();
-            ComputeNode worker = coordinatorPreprocessor.getWorkerProvider().getWorkerById(workerId);
-            TNetworkAddress execBeAddr = worker.getAddress();
-            receiver = new ResultReceiver(
-                    rootExecFragment.getInstances().get(0).getInstanceId(),
-                    workerId,
-                    worker.getBrpcAddress(),
-                    jobSpec.getQueryOptions().query_timeout * 1000);
+        boolean isLoadType = !(rootExecFragment.getPlanFragment().getSink() instanceof ResultSink);
+        if (isLoadType) {
+            return;
+        }
 
-            // Select top fragment as global runtime filter merge address
-            setGlobalRuntimeFilterParams(rootExecFragment, worker.getBrpcAddress());
+        long workerId = rootExecFragment.getInstances().get(0).getWorkerId();
+        ComputeNode worker = coordinatorPreprocessor.getWorkerProvider().getWorkerById(workerId);
+        TNetworkAddress execBeAddr = worker.getAddress();
+        receiver = new ResultReceiver(
+                rootExecFragment.getInstances().get(0).getInstanceId(),
+                workerId,
+                worker.getBrpcAddress(),
+                jobSpec.getQueryOptions().query_timeout * 1000);
 
-            if (LOG.isDebugEnabled()) {
-                LOG.debug("dispatch query job: {} to {}", DebugUtil.printId(jobSpec.getQueryId()), execBeAddr);
-            }
+        // Select top fragment as global runtime filter merge address
+        setGlobalRuntimeFilterParams(rootExecFragment, worker.getBrpcAddress());
 
-            // set the broker address for OUTFILE sink
-            ResultSink resultSink = (ResultSink) rootExecFragment.getPlanFragment().getSink();
-            if (resultSink.isOutputFileSink() && resultSink.needBroker()) {
-                FsBroker broker = GlobalStateMgr.getCurrentState().getBrokerMgr().getBroker(resultSink.getBrokerName(),
-                        execBeAddr.getHostname());
-                resultSink.setBrokerAddr(broker.ip, broker.port);
-                LOG.info("OUTFILE through broker: {}:{}", broker.ip, broker.port);
-            }
-        } else {
-            // This is a load process.
-            jobSpec.getQueryOptions().setEnable_profile(true);
-            deltaUrls = Lists.newArrayList();
-            loadCounters = Maps.newHashMap();
-            List<Long> relatedBackendIds = coordinatorPreprocessor.getWorkerProvider().getSelectedWorkerIds();
-            GlobalStateMgr.getCurrentState().getLoadMgr()
-                    .initJobProgress(jobSpec.getLoadJobId(), jobSpec.getQueryId(),
-                            executionDAG.getInstanceIds(),
-                            relatedBackendIds);
-            LOG.info("dispatch load job: {} to {}", DebugUtil.printId(jobSpec.getQueryId()),
-                    coordinatorPreprocessor.getWorkerProvider().getSelectedWorkerIds());
+        if (LOG.isDebugEnabled()) {
+            LOG.debug("dispatch query job: {} to {}", DebugUtil.printId(jobSpec.getQueryId()), execBeAddr);
+        }
+
+        // set the broker address for OUTFILE sink
+        ResultSink resultSink = (ResultSink) rootExecFragment.getPlanFragment().getSink();
+        if (resultSink.isOutputFileSink() && resultSink.needBroker()) {
+            FsBroker broker = GlobalStateMgr.getCurrentState().getBrokerMgr().getBroker(resultSink.getBrokerName(),
+                    execBeAddr.getHostname());
+            resultSink.setBrokerAddr(broker.ip, broker.port);
+            LOG.info("OUTFILE through broker: {}:{}", broker.ip, broker.port);
         }
     }
 
@@ -571,12 +527,7 @@ public class DefaultCoordinator extends Coordinator {
                 deployer.deployFragments(concurrentFragments, needDeploy);
             }
 
-            attachInstanceProfileToFragmentProfile();
-
-            // Avoid `explain scheduler` waits until profile timeout.
-            if (!needDeploy) {
-                profileDoneSignal.countDownToZero(Status.OK);
-            }
+            queryProfile.attachExecutionProfiles(executionDAG.getExecutions());
         } finally {
             unlock();
         }
@@ -695,11 +646,6 @@ public class DefaultCoordinator extends Coordinator {
     }
 
     @Override
-    public List<String> getExportFiles() {
-        return exportFiles;
-    }
-
-    @Override
     public Map<Integer, TNetworkAddress> getChannelIdToBEHTTPMap() {
         return coordinatorPreprocessor.getChannelIdToBEHTTPMap();
     }
@@ -707,97 +653,6 @@ public class DefaultCoordinator extends Coordinator {
     @Override
     public Map<Integer, TNetworkAddress> getChannelIdToBEPortMap() {
         return coordinatorPreprocessor.getChannelIdToBEPortMap();
-    }
-
-    void updateExportFiles(List<String> files) {
-        lock.lock();
-        try {
-            if (exportFiles == null) {
-                exportFiles = Lists.newArrayList();
-            }
-            exportFiles.addAll(files);
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    void updateDeltas(List<String> urls) {
-        lock.lock();
-        try {
-            deltaUrls.addAll(urls);
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    private void updateLoadCounters(Map<String, String> newLoadCounters) {
-        lock.lock();
-        try {
-            long numRowsNormal = 0L;
-            String value = this.loadCounters.get(LoadEtlTask.DPP_NORMAL_ALL);
-            if (value != null) {
-                numRowsNormal = Long.parseLong(value);
-            }
-            long numRowsAbnormal = 0L;
-            value = this.loadCounters.get(LoadEtlTask.DPP_ABNORMAL_ALL);
-            if (value != null) {
-                numRowsAbnormal = Long.parseLong(value);
-            }
-            long numRowsUnselected = 0L;
-            value = this.loadCounters.get(LoadJob.UNSELECTED_ROWS);
-            if (value != null) {
-                numRowsUnselected = Long.parseLong(value);
-            }
-            long numLoadBytesTotal = 0L;
-            value = this.loadCounters.get(LoadJob.LOADED_BYTES);
-            if (value != null) {
-                numLoadBytesTotal = Long.parseLong(value);
-            }
-
-            // new load counters
-            value = newLoadCounters.get(LoadEtlTask.DPP_NORMAL_ALL);
-            if (value != null) {
-                numRowsNormal += Long.parseLong(value);
-            }
-            value = newLoadCounters.get(LoadEtlTask.DPP_ABNORMAL_ALL);
-            if (value != null) {
-                numRowsAbnormal += Long.parseLong(value);
-            }
-            value = newLoadCounters.get(LoadJob.UNSELECTED_ROWS);
-            if (value != null) {
-                numRowsUnselected += Long.parseLong(value);
-            }
-            value = newLoadCounters.get(LoadJob.LOADED_BYTES);
-            if (value != null) {
-                numLoadBytesTotal += Long.parseLong(value);
-            }
-
-            this.loadCounters.put(LoadEtlTask.DPP_NORMAL_ALL, String.valueOf(numRowsNormal));
-            this.loadCounters.put(LoadEtlTask.DPP_ABNORMAL_ALL, String.valueOf(numRowsAbnormal));
-            this.loadCounters.put(LoadJob.UNSELECTED_ROWS, String.valueOf(numRowsUnselected));
-            this.loadCounters.put(LoadJob.LOADED_BYTES, String.valueOf(numLoadBytesTotal));
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    private void updateCommitInfos(List<TTabletCommitInfo> commitInfos) {
-        lock.lock();
-        try {
-            this.commitInfos.addAll(commitInfos);
-        } finally {
-            lock.unlock();
-        }
-    }
-
-    private void updateFailInfos(List<TTabletFailInfo> failInfos) {
-        lock.lock();
-        try {
-            this.failInfos.addAll(failInfos);
-            LOG.info(failInfos);
-        } finally {
-            lock.unlock();
-        }
     }
 
     private void updateStatus(Status status, TUniqueId instanceId) {
@@ -896,9 +751,10 @@ public class DefaultCoordinator extends Coordinator {
         return resultBatch;
     }
 
-    // Cancel execState of query. This includes the execState of the local plan
-    // fragment,
-    // if any, as well as all plan fragments on remote nodes.
+    /**
+     * Cancel execState of query. This includes the execState of the local plan fragment,
+     * if any, as well as all plan fragments on remote nodes.
+     */
     @Override
     public void cancel(PPlanFragmentCancelReason reason, String message) {
         lock();
@@ -916,9 +772,9 @@ public class DefaultCoordinator extends Coordinator {
             try {
                 // when enable_profile is true, it disable count down profileDoneSignal for collect all backend's profile
                 // but if backend has crashed, we need count down profileDoneSignal since it will not report by itself
-                if (connectContext.getSessionVariable().isEnableProfile() && profileDoneSignal != null
-                        && message.equals(FeConstants.BACKEND_NODE_NOT_FOUND_ERROR)) {
-                    profileDoneSignal.countDownToZero(new Status());
+                if (connectContext.getSessionVariable().isEnableProfile() &&
+                        message.equals(FeConstants.BACKEND_NODE_NOT_FOUND_ERROR)) {
+                    queryProfile.finishAllInstances(Status.OK);
                     LOG.info("count down profileDoneSignal since backend has crashed, query id: {}",
                             DebugUtil.printId(jobSpec.getQueryId()));
                 }
@@ -937,12 +793,10 @@ public class DefaultCoordinator extends Coordinator {
             receiver.cancel();
         }
         cancelRemoteFragmentsAsync(cancelReason);
-        if (profileDoneSignal != null && cancelReason != PPlanFragmentCancelReason.LIMIT_REACH) {
+        if (cancelReason != PPlanFragmentCancelReason.LIMIT_REACH) {
             // count down to zero to notify all objects waiting for this
             if (!connectContext.getSessionVariable().isEnableProfile()) {
-                profileDoneSignal.countDownToZero(new Status());
-                LOG.info("unfinished instance: {}",
-                        profileDoneSignal.getLeftMarks().stream().map(e -> DebugUtil.printId(e.getKey())).toArray());
+                queryProfile.finishAllInstances(Status.OK);
             }
         }
     }
@@ -954,13 +808,13 @@ public class DefaultCoordinator extends Coordinator {
             // because the profile report will not arrive anymore for the finished or non-deployed execState.
             if (!execState.cancelFragmentInstance(cancelReason) &&
                     (!execState.hasBeenDeployed() || execState.isFinished())) {
-                profileDoneSignal.markedCountDown(execState.getInstanceId(), -1L);
+                queryProfile.finishInstance(execState.getInstanceId());
             }
         }
 
         executionDAG.getInstances().stream()
                 .filter(instance -> executionDAG.getExecution(instance.getIndexInJob()) == null)
-                .forEach(instance -> profileDoneSignal.markedCountDown(instance.getInstanceId(), -1L));
+                .forEach(instance -> queryProfile.finishInstance(instance.getInstanceId()));
     }
 
     @Override
@@ -972,37 +826,7 @@ public class DefaultCoordinator extends Coordinator {
             return;
         }
 
-        if (params.isSetProfile()) {
-            profileAlreadyReported = true;
-        }
-
-        // Update runtime profile when query is still in process.
-        //
-        // We need to export profile to ProfileManager before update this profile, because:
-        // Each fragment instance will report its state based on their on own timer, and basically, these
-        // timers are consistent. So we can assume that all the instances will report profile in a very short
-        // time range, if we choose to export the profile to profile manager after update this instance's profile,
-        // the whole profile may include the information from the previous report, except for the current instance,
-        // which leads to inconsistency.
-        //
-        // So the profile update strategy looks like this: During a short time interval, each instance will report
-        // its execState information. However, when receiving the information reported by the first instance of the
-        // current batch, the previous reported state will be synchronized to the profile manager.
-        long now = System.currentTimeMillis();
-        long lastTime = lastRuntimeProfileUpdateTime.get();
-        if (topProfileSupplier != null && execPlanSupplier != null && connectContext != null &&
-                connectContext.getSessionVariable().isEnableProfile() &&
-                // If it's the last done report, avoiding duplicate trigger
-                (!execState.isFinished() || profileDoneSignal.getLeftMarks().size() > 1) &&
-                // Interval * 0.95 * 1000 to allow a certain range of deviation
-                now - lastTime > (connectContext.getSessionVariable().getRuntimeProfileReportInterval() * 950L) &&
-                lastRuntimeProfileUpdateTime.compareAndSet(lastTime, now)) {
-            RuntimeProfile profile = topProfileSupplier.get();
-            ExecPlan plan = execPlanSupplier.get();
-            profile.addChild(buildMergedQueryProfile());
-            ProfilingExecPlan profilingPlan = plan == null ? null : plan.getProfilingPlan();
-            ProfileManager.getInstance().pushProfile(profilingPlan, profile);
-        }
+        queryProfile.updateProfile(execState, params);
 
         lock();
         try {
@@ -1037,34 +861,22 @@ public class DefaultCoordinator extends Coordinator {
                     DebugUtil.printId(params.getFragment_instance_id()));
             updateStatus(status, params.getFragment_instance_id());
         }
+
         if (execState.isFinished()) {
-            if (params.isSetDelta_urls()) {
-                updateDeltas(params.getDelta_urls());
+            lock();
+            try {
+                queryProfile.updateLoadInformation(execState, params);
+            } finally {
+                unlock();
             }
-            if (params.isSetLoad_counters()) {
-                updateLoadCounters(params.getLoad_counters());
-            }
-            if (params.isSetTracking_url()) {
-                trackingUrl = params.tracking_url;
-            }
-            if (params.isSetExport_files()) {
-                updateExportFiles(params.export_files);
-            }
-            if (params.isSetCommitInfos()) {
-                updateCommitInfos(params.getCommitInfos());
-            }
-            if (params.isSetFailInfos()) {
-                updateFailInfos(params.getFailInfos());
-            }
-            if (params.isSetRejected_record_path()) {
-                rejectedRecordPaths.add(execState.getAddress().hostname + ":" + params.getRejected_record_path());
-            }
-            if (params.isSetSink_commit_infos()) {
-                sinkCommitInfos.addAll(params.sink_commit_infos);
-            }
-            profileDoneSignal.markedCountDown(params.getFragment_instance_id(), -1L);
+
+            queryProfile.finishInstance(params.getFragment_instance_id());
         }
 
+        updateJobProgress(params);
+    }
+
+    private void updateJobProgress(TReportExecStatusParams params) {
         if (params.isSetLoad_type()) {
             TLoadJobType loadJobType = params.getLoad_type();
             if (loadJobType == TLoadJobType.BROKER ||
@@ -1092,40 +904,34 @@ public class DefaultCoordinator extends Coordinator {
 
         // wait for all backends
         if (needReport) {
-            try {
-                int timeout;
-                // connectContext can be null for broker export task coordinator
-                if (connectContext != null) {
-                    timeout = connectContext.getSessionVariable().getProfileTimeout();
-                } else {
-                    timeout = DEFAULT_PROFILE_TIMEOUT_SECOND;
-                }
-                // Waiting for other fragment instances to finish execState
-                // Ideally, it should wait indefinitely, but out of defense, set timeout
-                if (!profileDoneSignal.await(timeout, TimeUnit.SECONDS)) {
-                    LOG.warn("failed to get profile within {} seconds", timeout);
-                }
-            } catch (InterruptedException e) { // NOSONAR
-                LOG.warn("signal await error", e);
+            int timeout;
+            // connectContext can be null for broker export task coordinator
+            if (connectContext != null) {
+                timeout = connectContext.getSessionVariable().getProfileTimeout();
+            } else {
+                timeout = DEFAULT_PROFILE_TIMEOUT_SECOND;
             }
+
+            // Waiting for other fragment instances to finish execState
+            // Ideally, it should wait indefinitely, but out of defense, set timeout
+            queryProfile.waitForProfileFinished(timeout, TimeUnit.SECONDS);
         }
+
         lock();
         try {
-            for (int i = 1; i < fragmentProfiles.size(); ++i) {
-                fragmentProfiles.get(i).sortChildren();
-            }
+            queryProfile.finalizeProfile();
         } finally {
             unlock();
         }
     }
 
-    /*
+    /**
      * Waiting the coordinator finish executing.
      * return false if waiting timeout.
      * return true otherwise.
      * NOTICE: return true does not mean that coordinator executed success,
      * the caller should check queryStatus for result.
-     *
+     * <p>
      * We divide the entire waiting process into multiple rounds,
      * with a maximum of 5 seconds per round. And after each round of waiting,
      * check the status of the BE. If the BE status is abnormal, the wait is ended
@@ -1140,12 +946,7 @@ public class DefaultCoordinator extends Coordinator {
         long leftTimeoutS = timeoutS;
         while (leftTimeoutS > 0) {
             long waitTime = Math.min(leftTimeoutS, fixedMaxWaitTime);
-            boolean awaitRes = false;
-            try {
-                awaitRes = profileDoneSignal.await(waitTime, TimeUnit.SECONDS);
-            } catch (InterruptedException e) { // NOSONAR
-                // Do nothing
-            }
+            boolean awaitRes = queryProfile.waitForProfileFinished(waitTime, TimeUnit.SECONDS);
             if (awaitRes) {
                 return true;
             }
@@ -1166,235 +967,7 @@ public class DefaultCoordinator extends Coordinator {
 
     @Override
     public RuntimeProfile buildMergedQueryProfile() {
-        SessionVariable sessionVariable = connectContext.getSessionVariable();
-
-        if (!sessionVariable.isEnableProfile()) {
-            return queryProfile;
-        }
-
-        if (!jobSpec.isEnablePipeline()) {
-            return queryProfile;
-        }
-
-        int profileLevel = sessionVariable.getPipelineProfileLevel();
-        if (profileLevel >= TPipelineProfileLevel.DETAIL.getValue()) {
-            return queryProfile;
-        }
-
-        RuntimeProfile newQueryProfile = new RuntimeProfile(queryProfile.getName());
-        newQueryProfile.copyAllInfoStringsFrom(queryProfile, null);
-        newQueryProfile.copyAllCountersFrom(queryProfile);
-
-        long maxQueryCumulativeCpuTime = 0;
-        long maxQueryPeakMemoryUsage = 0;
-        long maxQuerySpillBytes = 0;
-
-        List<RuntimeProfile> newFragmentProfiles = Lists.newArrayList();
-        for (RuntimeProfile fragmentProfile : fragmentProfiles) {
-            RuntimeProfile newFragmentProfile = new RuntimeProfile(fragmentProfile.getName());
-            newFragmentProfiles.add(newFragmentProfile);
-            newFragmentProfile.copyAllInfoStringsFrom(fragmentProfile, null);
-            newFragmentProfile.copyAllCountersFrom(fragmentProfile);
-
-            if (fragmentProfile.getChildList().isEmpty()) {
-                continue;
-            }
-
-            List<RuntimeProfile> instanceProfiles = fragmentProfile.getChildList().stream()
-                    .map(pair -> pair.first)
-                    .collect(Collectors.toList());
-
-            Set<String> backendAddresses = Sets.newHashSet();
-            Set<String> instanceIds = Sets.newHashSet();
-            Set<String> missingInstanceIds = Sets.newHashSet();
-            for (RuntimeProfile instanceProfile : instanceProfiles) {
-                // Setup backend meta infos
-                backendAddresses.add(instanceProfile.getInfoString("Address"));
-                instanceIds.add(instanceProfile.getInfoString("InstanceId"));
-                if (CollectionUtils.isEmpty(instanceProfile.getChildList())) {
-                    missingInstanceIds.add(instanceProfile.getInfoString("InstanceId"));
-                }
-
-                // Get query level peak memory usage and cpu cost
-                Counter toBeRemove = instanceProfile.getCounter("QueryCumulativeCpuTime");
-                if (toBeRemove != null) {
-                    maxQueryCumulativeCpuTime = Math.max(maxQueryCumulativeCpuTime, toBeRemove.getValue());
-                }
-                instanceProfile.removeCounter("QueryCumulativeCpuTime");
-
-                toBeRemove = instanceProfile.getCounter("QueryPeakMemoryUsage");
-                if (toBeRemove != null) {
-                    maxQueryPeakMemoryUsage = Math.max(maxQueryPeakMemoryUsage, toBeRemove.getValue());
-                }
-                instanceProfile.removeCounter("QueryPeakMemoryUsage");
-
-                toBeRemove = instanceProfile.getCounter("QuerySpillBytes");
-                if (toBeRemove != null) {
-                    maxQuerySpillBytes = Math.max(maxQuerySpillBytes, toBeRemove.getValue());
-                }
-                instanceProfile.removeCounter("QuerySpillBytes");
-            }
-            newFragmentProfile.addInfoString("BackendAddresses", String.join(",", backendAddresses));
-            newFragmentProfile.addInfoString("InstanceIds", String.join(",", instanceIds));
-            if (!missingInstanceIds.isEmpty()) {
-                newFragmentProfile.addInfoString("MissingInstanceIds", String.join(",", missingInstanceIds));
-            }
-            Counter backendNum = newFragmentProfile.addCounter("BackendNum", TUnit.UNIT, null);
-            backendNum.setValue(backendAddresses.size());
-
-            // Setup number of instance
-            Counter counter = newFragmentProfile.addCounter("InstanceNum", TUnit.UNIT, null);
-            counter.setValue(instanceProfiles.size());
-
-            RuntimeProfile mergedInstanceProfile =
-                    RuntimeProfile.mergeIsomorphicProfiles(instanceProfiles, Sets.newHashSet("Address", "InstanceId"));
-            Preconditions.checkState(mergedInstanceProfile != null);
-
-            newFragmentProfile.copyAllInfoStringsFrom(mergedInstanceProfile, null);
-            newFragmentProfile.copyAllCountersFrom(mergedInstanceProfile);
-
-            mergedInstanceProfile.getChildList().forEach(pair -> {
-                RuntimeProfile pipelineProfile = pair.first;
-                foldUnnecessaryLimitOperators(pipelineProfile);
-                setOperatorStatus(pipelineProfile);
-                newFragmentProfile.addChild(pipelineProfile);
-            });
-
-            newQueryProfile.addChild(newFragmentProfile);
-        }
-
-        // Remove redundant MIN/MAX metrics if MIN and MAX are identical
-        for (RuntimeProfile fragmentProfile : newFragmentProfiles) {
-            RuntimeProfile.removeRedundantMinMaxMetrics(fragmentProfile);
-        }
-
-        long queryAllocatedMemoryUsage = 0;
-        long queryDeallocatedMemoryUsage = 0;
-        // Calculate ExecutionTotalTime, which comprising all operator's sync time and async time
-        // We can get Operator's sync time from OperatorTotalTime, and for async time, only ScanOperator and
-        // ExchangeOperator have async operations, we can get async time from ScanTime(for ScanOperator) and
-        // NetworkTime(for ExchangeOperator)
-        long queryCumulativeOperatorTime = 0;
-        boolean foundResultSink = false;
-        for (RuntimeProfile fragmentProfile : newFragmentProfiles) {
-            Counter instanceAllocatedMemoryUsage = fragmentProfile.getCounter("InstanceAllocatedMemoryUsage");
-            if (instanceAllocatedMemoryUsage != null) {
-                queryAllocatedMemoryUsage += instanceAllocatedMemoryUsage.getValue();
-            }
-            Counter instanceDeallocatedMemoryUsage = fragmentProfile.getCounter("InstanceDeallocatedMemoryUsage");
-            if (instanceDeallocatedMemoryUsage != null) {
-                queryDeallocatedMemoryUsage += instanceDeallocatedMemoryUsage.getValue();
-            }
-
-            for (Pair<RuntimeProfile, Boolean> pipelineProfilePair : fragmentProfile.getChildList()) {
-                RuntimeProfile pipelineProfile = pipelineProfilePair.first;
-                for (Pair<RuntimeProfile, Boolean> operatorProfilePair : pipelineProfile.getChildList()) {
-                    RuntimeProfile operatorProfile = operatorProfilePair.first;
-                    if (!foundResultSink && (operatorProfile.getName().contains("RESULT_SINK") ||
-                            operatorProfile.getName().contains("OLAP_TABLE_SINK"))) {
-                        newQueryProfile.getCounterTotalTime().setValue(0);
-
-                        long executionWallTime = pipelineProfile.getCounter("DriverTotalTime").getValue();
-                        Counter executionTotalTime =
-                                newQueryProfile.addCounter("QueryExecutionWallTime", TUnit.TIME_NS, null);
-                        executionTotalTime.setValue(executionWallTime);
-
-                        Counter outputFullTime = pipelineProfile.getCounter("OutputFullTime");
-                        if (outputFullTime != null) {
-                            long resultDeliverTime = outputFullTime.getValue();
-                            Counter resultDeliverTimer =
-                                    newQueryProfile.addCounter("ResultDeliverTime", TUnit.TIME_NS, null);
-                            resultDeliverTimer.setValue(resultDeliverTime);
-                        }
-
-                        foundResultSink = true;
-                    }
-
-                    RuntimeProfile commonMetrics = operatorProfile.getChild("CommonMetrics");
-                    RuntimeProfile uniqueMetrics = operatorProfile.getChild("UniqueMetrics");
-                    if (commonMetrics == null || uniqueMetrics == null) {
-                        continue;
-                    }
-                    Counter operatorTotalTime = commonMetrics.getMaxCounter("OperatorTotalTime");
-                    Preconditions.checkNotNull(operatorTotalTime);
-                    queryCumulativeOperatorTime += operatorTotalTime.getValue();
-
-                    Counter scanTime = uniqueMetrics.getMaxCounter("ScanTime");
-                    if (scanTime != null) {
-                        queryCumulativeOperatorTime += scanTime.getValue();
-                    }
-
-                    Counter networkTime = uniqueMetrics.getMaxCounter("NetworkTime");
-                    if (networkTime != null) {
-                        queryCumulativeOperatorTime += networkTime.getValue();
-                    }
-                }
-            }
-        }
-        Counter queryAllocatedMemoryUsageCounter =
-                newQueryProfile.addCounter("QueryAllocatedMemoryUsage", TUnit.BYTES, null);
-        queryAllocatedMemoryUsageCounter.setValue(queryAllocatedMemoryUsage);
-        Counter queryDeallocatedMemoryUsageCounter =
-                newQueryProfile.addCounter("QueryDeallocatedMemoryUsage", TUnit.BYTES, null);
-        queryDeallocatedMemoryUsageCounter.setValue(queryDeallocatedMemoryUsage);
-        Counter queryCumulativeOperatorTimer =
-                newQueryProfile.addCounter("QueryCumulativeOperatorTime", TUnit.TIME_NS, null);
-        queryCumulativeOperatorTimer.setValue(queryCumulativeOperatorTime);
-        newQueryProfile.getCounterTotalTime().setValue(0);
-
-        Counter queryCumulativeCpuTime = newQueryProfile.addCounter("QueryCumulativeCpuTime", TUnit.TIME_NS, null);
-        queryCumulativeCpuTime.setValue(maxQueryCumulativeCpuTime);
-        Counter queryPeakMemoryUsage = newQueryProfile.addCounter("QueryPeakMemoryUsage", TUnit.BYTES, null);
-        queryPeakMemoryUsage.setValue(maxQueryPeakMemoryUsage);
-        Counter querySpillBytes = newQueryProfile.addCounter("QuerySpillBytes", TUnit.BYTES, null);
-        querySpillBytes.setValue(maxQuerySpillBytes);
-
-        return newQueryProfile;
-    }
-
-    /**
-     * Remove unnecessary LimitOperator, which has same input rows and output rows
-     * to keep the profile concise
-     */
-    private void foldUnnecessaryLimitOperators(RuntimeProfile pipelineProfile) {
-        SessionVariable sessionVariable = connectContext.getSessionVariable();
-        if (!sessionVariable.isProfileLimitFold()) {
-            return;
-        }
-
-        List<String> foldNames = Lists.newArrayList();
-        for (Pair<RuntimeProfile, Boolean> child : pipelineProfile.getChildList()) {
-            RuntimeProfile operatorProfile = child.first;
-            if (operatorProfile.getName().contains("LIMIT")) {
-                RuntimeProfile commonMetrics = operatorProfile.getChild("CommonMetrics");
-                Preconditions.checkNotNull(commonMetrics);
-                Counter pullRowNum = commonMetrics.getCounter("PullRowNum");
-                Counter pushRowNum = commonMetrics.getCounter("PushRowNum");
-                if (pullRowNum == null || pushRowNum == null) {
-                    continue;
-                }
-                if (Objects.equals(pullRowNum.getValue(), pushRowNum.getValue())) {
-                    foldNames.add(operatorProfile.getName());
-                }
-            }
-        }
-
-        foldNames.forEach(pipelineProfile::removeChild);
-    }
-
-    private void setOperatorStatus(RuntimeProfile pipelineProfile) {
-        for (Pair<RuntimeProfile, Boolean> child : pipelineProfile.getChildList()) {
-            RuntimeProfile operatorProfile = child.first;
-            RuntimeProfile commonMetrics = operatorProfile.getChild("CommonMetrics");
-            Preconditions.checkNotNull(commonMetrics);
-
-            Counter closeTime = commonMetrics.getCounter("CloseTime");
-            Counter minCloseTime = commonMetrics.getCounter("__MIN_OF_CloseTime");
-            if (closeTime != null && closeTime.getValue() == 0 ||
-                    minCloseTime != null && minCloseTime.getValue() == 0) {
-                commonMetrics.addInfoString("Status", "Running");
-            }
-        }
+        return queryProfile.buildMergedQueryProfile();
     }
 
     /**
@@ -1415,7 +988,7 @@ public class DefaultCoordinator extends Coordinator {
 
     @Override
     public boolean isDone() {
-        return profileDoneSignal.getCount() == 0;
+        return queryProfile.isFinished();
     }
 
     @Override
@@ -1423,19 +996,14 @@ public class DefaultCoordinator extends Coordinator {
         return connectContext != null && connectContext.getSessionVariable().isEnableLoadProfile();
     }
 
-    // consistent with EXPLAIN's fragment index
+    /**
+     * Get information of each fragment instance.
+     *
+     * @return the fragment instance information list, consistent with the fragment index of {@code EXPLAIN}.
+     */
     @Override
     public List<QueryStatisticsItem.FragmentInstanceInfo> getFragmentInstanceInfos() {
         return executionDAG.getFragmentInstanceInfos();
-    }
-
-    private void attachInstanceProfileToFragmentProfile() {
-        for (FragmentInstanceExecState execState : executionDAG.getExecutions()) {
-            if (!execState.computeTimeInProfile(fragmentProfiles.size())) {
-                return;
-            }
-            fragmentProfiles.get(execState.getFragmentIndex()).addChild(execState.getProfile());
-        }
     }
 
     @Override
@@ -1445,6 +1013,6 @@ public class DefaultCoordinator extends Coordinator {
 
     @Override
     public boolean isProfileAlreadyReported() {
-        return this.profileAlreadyReported;
+        return this.queryProfile.isProfileAlreadyReported();
     }
 }
